refactor: 리뷰더보기 버튼에 뜨는  문구 수정 및 남은  개수출력
