+++ src/features/placeDetail/PlaceDetailContainer.tsx
@@ -1,19 +1,21 @@
 import PhotoStrip from './PhotoStrip';
 import PlaceInfo from './PlaceInfo';
 import { useParams } from 'react-router-dom';
-import ReviewStats from './ReviewStats';
+import ReviewStatsSection from './ReviewStatsSection';
 import ReviewCard from './place-review-card/ReviewCard';
 import ReviewWriteButton from './ReviewWriteButton';
 import MoreReviewButton from './MoreReviewButton';
 import { usePlaceDetail } from '../../hooks/usePlaceDetail';
 import { usePlaceReview } from '../../hooks/usePlaceReview';
+import { useReviewStats } from '../../hooks/useReviewStats';
 
 const PlaceDetailContainer = () => {
   const { id } = useParams<{ id: string }>();
   const { place } = usePlaceDetail(id!);
   const { reviews } = usePlaceReview(id!);
+  const { stats } = useReviewStats(id!);
 
-  if (!place) return <div>로딩중...</div>;
+  if (!place || !stats) return <div>로딩중...</div>;
 
   return (
     <div className="mx-auto mt-12 flex w-[75%] flex-col items-center gap-10 overflow-y-auto">
@@ -22,7 +24,7 @@ const PlaceDetailContainer = () => {
       <div className="flex w-full border border-gray-100" />
       <ReviewWriteButton placeName={place.name} placeId={id!} />
       <div className="flex w-full border border-gray-100" />
-      <ReviewStats />
+      <ReviewStatsSection stats={stats} />
       <div className="flex w-[90%] flex-col">
         {reviews.slice(0, 2).map((review) => (
           <>
@@ -30,7 +32,9 @@ const PlaceDetailContainer = () => {
             <ReviewCard key={review.reviewId} review={review} placeId={id!} />
           </>
         ))}
-        {reviews[2] && <MoreReviewButton />}
+        {reviews.length > 2 && (
+          <MoreReviewButton reviewCount={stats.reviewCount} />
+        )}
       </div>
     </div>
   );
